// src/lib/api/auth.ts
export interface AuthResponse {
  accessToken: string;
  expiresIn: number;
  refreshToken: string;
  tokenType: string;
}

export interface ForgotPasswordRequest {
  email: string;
  /** @nullable */
  password?: string | null;
}

export interface LicenseStatusResponse {
  /** @nullable */
  expiresAt?: string | null;
  hasLicense: boolean;
  isValid: boolean;
  /** @nullable */
  licenseKey?: string | null;
  /** @nullable */
  subscriptionType?: string | null;
}
// ...
export interface ProfileResponse {
  /** @nullable */
  company?: string | null;
  email: string;
  emailVerified: boolean;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  userId: string;
}